+++ src/cross_sections.c
@@ -26,7 +26,7 @@ THIS SOFTWARE IS PROVIDED BY Bruno Golosio, Antonio Brunetti, Manuel Sanchez del
 //          E : energy (keV)                                        //
 //                                                                  //
 /////////////////////////////////////////////////////////////////// */
-float CS_Total(int Z, float E)
+double CS_Total(int Z, double E)
 {
   if (Z<1 || Z>ZMAX || NE_Photo[Z]<0 || NE_Rayl[Z]<0 || NE_Compt[Z]<0) {
     ErrorExit("Z out of range in function CS_Total");
@@ -49,9 +49,9 @@ float CS_Total(int Z, float E)
 //          E : energy (keV)                                        //
 //                                                                  //
 /////////////////////////////////////////////////////////////////// */
-float CS_Photo(int Z, float E)
+double CS_Photo(int Z, double E)
 {
-  float ln_E, ln_sigma, sigma;
+  double ln_E, ln_sigma, sigma;
 
   if (Z<1 || Z>ZMAX || NE_Photo[Z]<0) {
     ErrorExit("Z out of range in function CS_Photo");
@@ -81,9 +81,9 @@ float CS_Photo(int Z, float E)
 //          E : energy (keV)                                        //
 //                                                                  //
 /////////////////////////////////////////////////////////////////// */
-float CS_Rayl(int Z, float E)
+double CS_Rayl(int Z, double E)
 {
-  float ln_E, ln_sigma, sigma;
+  double ln_E, ln_sigma, sigma;
 
   if (Z<1 || Z>ZMAX || NE_Rayl[Z]<0) {
     ErrorExit("Z out of range in function CS_Rayl");
@@ -112,9 +112,9 @@ float CS_Rayl(int Z, float E)
 //          E : energy (keV)                                        //
 //                                                                  //
 /////////////////////////////////////////////////////////////////// */
-float CS_Compt(int Z, float E) 
+double CS_Compt(int Z, double E) 
 {
-  float ln_E, ln_sigma, sigma;
+  double ln_E, ln_sigma, sigma;
 
   if (Z<1 || Z>ZMAX || NE_Compt[Z]<0) {
     ErrorExit("Z out of range in function CS_Compt");
@@ -145,9 +145,9 @@ float CS_Compt(int Z, float E)
 //          E : energy (keV)                                        //
 //                                                                  //
 /////////////////////////////////////////////////////////////////// */
-float CS_Energy(int Z, float E)
+double CS_Energy(int Z, double E)
 {
-	float ln_E, ln_sigma, sigma;
+	double ln_E, ln_sigma, sigma;
 	if (Z < 1 || Z > 92 || NE_Energy[Z] < 0) {
 		ErrorExit("Z out of range in function CS_Energy");
 		return 0;
